FE-Refactor: Sidepanel 관련 로직 sidepanel 폴더 하위로 이동
